2024.12.31 CODEKATA : 26 ~ 30번

+++ src/Programmers/Codekata/_23_콜라츠추측.java
@@ -3,7 +3,7 @@ package Programmers.Codekata;
 public class _23_콜라츠추측 {
     class Solution {
         public int solution(int num) {
-            long n = (long)num;
+            long n = num;
             int answer = 0;
             while(answer <= 500){
                 if(n == 1) break;
@@ -13,7 +13,6 @@ public class _23_콜라츠추측 {
                 }else{
                     n = n * 3 + 1;
                 }
-
                 answer++;
             }
             return answer > 500 ? -1 : answer;
